move getFilterLabel to its own file;
+ add cases for GeoPoint and DateArray
